refactor: スキップボタンをコンポーネント
